+++ files/chef-marketplace-cookbooks/chef-marketplace/recipes/_analytics_preconfigure.rb
@@ -1,5 +1,6 @@
 bash 'opscode-analytics-ctl reconfigure' do
   code 'opscode-analytics-ctl reconfigure'
+  live_stream true
 end
 
 bash 'opscode-analytics-ctl stop' do